+++ icenet_mp/models/common/__init__.py
@@ -2,13 +2,14 @@ from .conv_block_common import CommonConvBlock
 from .conv_block_downsample import ConvBlockDownsample
 from .conv_block_upsample import ConvBlockUpsample
 from .conv_norm_act_upsample import ConvNormActUpsample
-from .dcae_blocks import ResBlock, SelfAttention2D
 from .mask import Mask
 from .normalised_fold import NormalisedFold
 from .patchembed import PatchEmbedding
 from .permute import Permute
+from .res_block import ResBlock
 from .resizing_interpolation import ResizingInterpolation
 from .restrict_range import RestrictRange
+from .self_attention import SelfAttention2D
 from .shift import Shift
 from .time_embed import TimeEmbed
 from .transformerblock import TransformerEncoderBlock
